Move Lance commit helpers out of writer
