+++ nemo_curator/stages/text/io/writer/__init__.py
@@ -12,13 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from nemo_curator.stages.text.io.writer.jsonl import JsonlWriter
-from nemo_curator.stages.text.io.writer.lance import (
-    LanceAnnotationWriter,
-    LanceWriter,
+from nemo_curator.stages.text.io.lance_commit import (
     commit_lance_annotation_checkpoint,
     commit_lance_checkpoint,
 )
+from nemo_curator.stages.text.io.writer.jsonl import JsonlWriter
+from nemo_curator.stages.text.io.writer.lance import LanceAnnotationWriter, LanceWriter
 from nemo_curator.stages.text.io.writer.parquet import ParquetWriter
 
 __all__ = [
